Extract classpath factory methods into separate trait.
Extracted methods responsible for classpath element creation into
separate trait that can be implemented to create either recursive
(old classpath) or flat (new classpath) elements.
