+++ src/compiler/scala/tools/util/PathResolver.scala
@@ -15,6 +15,7 @@ import scala.reflect.runtime.ReflectionUtils
 import ClassPath.{ JavaContext, DefaultJavaContext, join, split }
 import PartialFunction.condOpt
 import scala.language.postfixOps
+import scala.tools.nsc.classpath.ClasspathFactory
 
 // Loosely based on the draft specification at:
 // https://wiki.scala-lang.org/display/SIW/Classpath
@@ -197,21 +198,16 @@ object PathResolver {
   }
 }
 
-class PathResolver(settings: Settings, context: JavaContext) {
-  import PathResolver.{ Defaults, Environment, AsLines, MkLines, ppcp }
-
-  def this(settings: Settings) = this(settings,
-      if (settings.YnoLoadImplClass) PathResolver.NoImplClassJavaContext
-      else DefaultJavaContext)
-
-  private def cmdLineOrElse(name: String, alt: String) = {
+abstract class PathResolverBase[T](settings: Settings, classPathFactory: ClasspathFactory[T]) {
+  import PathResolver.{Defaults, ppcp, AsLines}
+  protected def cmdLineOrElse(name: String, alt: String) = {
     (commandLineFor(name) match {
       case Some("") => None
       case x        => x
     }) getOrElse alt
   }
 
-  private def commandLineFor(s: String): Option[String] = condOpt(s) {
+  protected def commandLineFor(s: String): Option[String] = condOpt(s) {
     case "javabootclasspath"  => settings.javabootclasspath.value
     case "javaextdirs"        => settings.javaextdirs.value
     case "bootclasspath"      => settings.bootclasspath.value
@@ -256,10 +252,10 @@ class PathResolver(settings: Settings, context: JavaContext) {
       else sys.env.getOrElse("CLASSPATH", ".")
     )
 
-    import context._
+    import classPathFactory._
 
     // Assemble the elements!
-    def basis = List[Traversable[ClassPath[AbstractFile]]](
+    def basis = List[Traversable[T]](
       classesInPath(javaBootClassPath),             // 1. The Java bootstrap class path.
       contentsOfDirsInPath(javaExtDirs),            // 2. The Java extension class path.
       classesInExpandedPath(javaUserClassPath),     // 3. The Java application class path.
@@ -285,6 +281,16 @@ class PathResolver(settings: Settings, context: JavaContext) {
       |  sourcePath           = ${ppcp(sourcePath)}
       |}""".asLines
   }
+}
+
+class PathResolver(settings: Settings, context: JavaContext) extends
+  PathResolverBase[ClassPath[AbstractFile]](settings, context) {
+  import PathResolver.{ Defaults, Environment, MkLines}
+  def this(settings: Settings) = this(settings, if (settings.inline) new JavaContext else DefaultJavaContext)
+
+  def this(settings: Settings) = this(settings,
+      if (settings.YnoLoadImplClass) PathResolver.NoImplClassJavaContext
+      else DefaultJavaContext)
 
   def containers = Calculated.containers
